refactor: store 명칭 변경
